Module vortex_flatbuffers::array

Structs

Array

ArrayArgs

ArrayBuilder

ArrayStats

ArrayStatsArgs

ArrayStatsBuilder

Version

Enums

ArrayOffset

ArrayStatsOffset

Constants

ENUM_MAX_VERSIONDeprecated

ENUM_MIN_VERSIONDeprecated

ENUM_VALUES_VERSIONDeprecated

Functions

finish_array_buffer

finish_size_prefixed_array_buffer

root_as_array

Verifies that a buffer of bytes contains a Array and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_array_unchecked.

root_as_array_unchecked⚠

Assumes, without verification, that a buffer of bytes contains a Array and returns it.

root_as_array_with_opts

Verifies, with the given options, that a buffer of bytes contains a Array and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_array_unchecked.

size_prefixed_root_as_array

Verifies that a buffer of bytes contains a size prefixed Array and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use size_prefixed_root_as_array_unchecked.

size_prefixed_root_as_array_unchecked⚠

Assumes, without verification, that a buffer of bytes contains a size prefixed Array and returns it.

size_prefixed_root_as_array_with_opts

Verifies, with the given verifier options, that a buffer of bytes contains a size prefixed Array and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_array_unchecked.