Module vortex_flatbuffers::message

Structs

Batch

BatchArgs

BatchBuilder

Buffer

Compression

Message

MessageArgs

MessageBuilder

MessageHeader

MessageHeaderUnionTableOffset

MessageVersion

Page

PageArgs

PageBuilder

Schema

SchemaArgs

SchemaBuilder

Enums

BatchOffset

MessageOffset

PageOffset

SchemaOffset

Constants

ENUM_MAX_COMPRESSIONDeprecated

ENUM_MAX_MESSAGE_HEADERDeprecated

ENUM_MAX_MESSAGE_VERSIONDeprecated

ENUM_MIN_COMPRESSIONDeprecated

ENUM_MIN_MESSAGE_HEADERDeprecated

ENUM_MIN_MESSAGE_VERSIONDeprecated

ENUM_VALUES_COMPRESSIONDeprecated

ENUM_VALUES_MESSAGE_HEADERDeprecated

ENUM_VALUES_MESSAGE_VERSIONDeprecated

Functions

finish_message_buffer

finish_size_prefixed_message_buffer

root_as_message

Verifies that a buffer of bytes contains a Message and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_message_unchecked.

root_as_message_unchecked⚠

Assumes, without verification, that a buffer of bytes contains a Message and returns it.

root_as_message_with_opts

Verifies, with the given options, that a buffer of bytes contains a Message and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_message_unchecked.

size_prefixed_root_as_message

Verifies that a buffer of bytes contains a size prefixed Message and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use size_prefixed_root_as_message_unchecked.

size_prefixed_root_as_message_unchecked⚠

Assumes, without verification, that a buffer of bytes contains a size prefixed Message and returns it.

size_prefixed_root_as_message_with_opts

Verifies, with the given verifier options, that a buffer of bytes contains a size prefixed Message and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_message_unchecked.