Module vortex_flatbuffers::scalar

Structs

Scalar

ScalarArgs

ScalarBuilder

ScalarValue

ScalarValueArgs

ScalarValueBuilder

Enums

ScalarOffset

ScalarValueOffset

Functions

finish_scalar_buffer

finish_size_prefixed_scalar_buffer

root_as_scalar

Verifies that a buffer of bytes contains a Scalar and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_scalar_unchecked.

root_as_scalar_unchecked⚠

Assumes, without verification, that a buffer of bytes contains a Scalar and returns it.

root_as_scalar_with_opts

Verifies, with the given options, that a buffer of bytes contains a Scalar and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_scalar_unchecked.

size_prefixed_root_as_scalar

Verifies that a buffer of bytes contains a size prefixed Scalar and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use size_prefixed_root_as_scalar_unchecked.

size_prefixed_root_as_scalar_unchecked⚠

Assumes, without verification, that a buffer of bytes contains a size prefixed Scalar and returns it.

size_prefixed_root_as_scalar_with_opts

Verifies, with the given verifier options, that a buffer of bytes contains a size prefixed Scalar and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_scalar_unchecked.